% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_date_cols}
\alias{get_date_cols}
\title{Convert specified columns to Date format}
\usage{
get_date_cols(df, date_cols)
}
\arguments{
\item{df}{The input data frame}

\item{date_cols}{Character vector of date column names}
}
\value{
Data frame with specified columns converted to Date format
}
\description{
Takes a data frame and converts specified date columns to Date format using the specified date format.
}
\examples{
\dontrun{
# Example usage:
data <- data.frame(date1 = c("01/01/2022", "02/01/2022"), date2 = c("03/01/2022", "04/01/2022"))
date_data <- get_date_cols(data, c("date1", "date2"))
}
}
