% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraper.R
\name{get_duration}
\alias{get_duration}
\title{Get Duration Data by Date Range}
\usage{
get_duration(
  start_date = Sys.Date() - 21,
  end_date = Sys.Date(),
  cvm_date = FALSE,
  cetip_code = NULL
)
}
\arguments{
\item{start_date}{The start date for retrieving duration data. Defaults to 21 days ago from the current date.}

\item{end_date}{The end date for retrieving duration data. Defaults to the current date.}

\item{cvm_date}{Logical indicating whether to use CVM date format.}

\item{cetip_code}{The CETIP code for the desired duration data. Defaults to NULL.}
}
\value{
A data frame containing the duration data for the specified date range and optional CETIP code.
}
\description{
This function retrieves duration data from the debentures.com.br website for a specified date range.
}
\examples{
# Retrieve duration data for the last 21 days
duration_data <- get_duration()

# Retrieve duration data for a specific date range and CETIP code
duration_data <- get_duration(start_date = "2022-01-01", end_date = "2022-12-31", cetip_code = "AALM11")

}
\keyword{data}
\keyword{debentures}
\keyword{duration}
\keyword{finance}
