% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraper.R
\name{get_issuing_prices}
\alias{get_issuing_prices}
\title{Get Issuing Prices Data by Date Range}
\usage{
get_issuing_prices(
  cetip_code,
  start_date = Sys.Date() - 5,
  end_date = Sys.Date()
)
}
\arguments{
\item{cetip_code}{The CETIP code for the desired storage data.}

\item{start_date}{The start date for retrieving issuing prices data. Defaults to 5 days ago from the current date.}

\item{end_date}{The end date for retrieving issuing prices data. Defaults to the current date.}
}
\value{
A data frame containing the issuing prices data for the specified date range.
}
\description{
This function retrieves issuing prices data from the debentures.com.br website for a specified date range.
}
\examples{
# Retrieve issuing prices data for the last week
issuing_prices_data <- get_issuing_prices()

# Retrieve issuing prices data for a specific date range
issuing_prices_data <- get_issuing_prices(cetip_code = 'AALM11', start_date = "2022-01-01", end_date = "2022-12-31")

}
\keyword{data}
\keyword{debentures}
\keyword{finance}
\keyword{issuing}
\keyword{prices}
