% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraper.R
\name{get_issuing_volume_by_date}
\alias{get_issuing_volume_by_date}
\title{Get Issuing Volume Data by Date Range}
\usage{
get_issuing_volume_by_date(
  start_date = Sys.Date() - 21,
  end_date = Sys.Date() - 1,
  date_type = "issuing",
  cvm_instruction = c(400, 476)
)
}
\arguments{
\item{start_date}{The start date for retrieving issuing volume data. Defaults to 21 days ago from the current date.}

\item{end_date}{The end date for retrieving issuing volume data. Defaults to 1 day ago from the current date.}

\item{date_type}{A character vector specifying the date type ('issuing', 'registering'). Defaults to 'issuing'.}

\item{cvm_instruction}{A numeric vector specifying the CVM instructions (400, 476). Defaults to c(400, 476).}
}
\value{
A data frame containing issuing volume data for the specified date range, date type, and CVM instructions.
}
\description{
This function retrieves issuing volume data from the debentures.com.br website for a specified date range.
}
\examples{
# Retrieve issuing volume data for the last 21 days
issuing_volume_data <- get_issuing_volume_by_date()

# Retrieve issuing volume data for a specific date range and CVM instructions
issuing_volume_data <- get_issuing_volume_by_date(start_date = "2022-01-01", end_date = "2022-12-31", date_type = "registering", cvm_instruction = c(400, 476))

}
\keyword{data}
\keyword{debentures}
\keyword{finance}
\keyword{issuing}
\keyword{volume}
