% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraper.R
\name{get_registered_bakers}
\alias{get_registered_bakers}
\title{Get Registered Banks (Bakers)}
\usage{
get_registered_bakers()
}
\value{
A data frame containing information about registered banks (bakers).
}
\description{
This function retrieves data about registered banks (bakers) from the Debentures website.
}
\examples{
# Retrieve information about registered banks (bakers)
banks_data <- get_registered_bakers()

}
\keyword{bakers}
\keyword{banks}
\keyword{data}
\keyword{debentures}
\keyword{finance}
