% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraper.R
\name{get_registered_bankers}
\alias{get_registered_bankers}
\title{Get Registered Banks (Bankers)}
\usage{
get_registered_bankers()
}
\value{
A data frame containing information about registered banks (bankers).
}
\description{
This function retrieves data about registered banks (bankers) from the debentures.com.br website.
}
\examples{
# Retrieve information about registered banks (bankers)
banks_data <- get_registered_bankers()

}
\keyword{bankers}
\keyword{banks}
\keyword{data}
\keyword{debentures}
\keyword{finance}
