% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraper.R
\name{get_registered_coordinators}
\alias{get_registered_coordinators}
\title{Get Registered Coordinators}
\usage{
get_registered_coordinators()
}
\value{
A data frame containing information about registered coordinators.
}
\description{
This function retrieves data about registered coordinators from the Debentures website.
}
\examples{
# Retrieve information about registered coordinators
coordinators_data <- get_registered_coordinators()

}
\keyword{coordinators}
\keyword{data}
\keyword{debentures}
\keyword{finance}
