% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraper.R
\name{get_registered_trustees}
\alias{get_registered_trustees}
\title{Get Registered Trustees}
\usage{
get_registered_trustees()
}
\value{
A data frame containing information about registered trustees.
}
\description{
This function retrieves data about registered trustees from the Debentures website.
}
\examples{
# Retrieve information about registered trustees
trustees_data <- get_registered_trustees()

}
\keyword{data}
\keyword{debentures}
\keyword{finance}
\keyword{trustees}
