% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_request_content}
\alias{get_request_content}
\title{Get content from an HTTP response object}
\usage{
get_request_content(res, encoding = "latin1")
}
\arguments{
\item{res}{The HTTP response object}

\item{encoding}{The character encoding to be used (default is "latin1")}
}
\value{
Extracted content
}
\description{
Extracts content from an HTTP response object and returns it.
}
\examples{
\dontrun{
# Example usage:
url <- "https://example.com/data"
response <- request_data(url)
content <- get_request_content(response)
}
}
