% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraper.R
\name{get_storage_by_cetip_code}
\alias{get_storage_by_cetip_code}
\title{Get Storage Data by CETIP Code and Date Range}
\usage{
get_storage_by_cetip_code(
  cetip_code,
  start_date = Sys.Date(),
  end_date = Sys.Date()
)
}
\arguments{
\item{cetip_code}{The CETIP code for the desired storage data.}

\item{start_date}{The start date for retrieving storage data. Defaults to the current date.}

\item{end_date}{The end date for retrieving storage data. Defaults to the current date.}
}
\value{
A data frame containing the storage data for the specified CETIP code and date range.
}
\description{
This function retrieves storage data from the Debentures website for a specified CETIP code and date range.
}
\examples{
# Retrieve storage data for a specific CETIP code and date range
storage_data <- get_storage_by_cetip_code(cetip_code = "ABC123", start_date = "2022-01-01", end_date = "2022-12-31")

}
\keyword{CETIP}
\keyword{code}
\keyword{data}
\keyword{debentures}
\keyword{finance}
\keyword{storage}
