% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraper.R
\name{get_storage_by_date}
\alias{get_storage_by_date}
\title{Get Storage Data by Date Range}
\usage{
get_storage_by_date(date = Sys.Date() - 5)
}
\arguments{
\item{date}{Date for retrieving storage data. Defaults to 5 days ago from the current date.}
}
\value{
A data frame containing the storage data for the specified date range.
}
\description{
This function retrieves storage data from the debentures.com.br website for a specified date range.
}
\examples{
# Retrieve storage data for the last week
storage_data <- get_storage_by_date()

# Retrieve storage data for a specific date
storage_data <- get_storage_by_date(date = "2022-01-03")

}
\keyword{data}
\keyword{debentures}
\keyword{finance}
\keyword{storage}
