% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraper.R
\name{get_storage_by_indexer}
\alias{get_storage_by_indexer}
\title{Get Storage Data by Indexer and Date Range}
\usage{
get_storage_by_indexer(
  indexer = c("Prefixado", "ANBID", "BTN", "DI", "DOLAR", "FDS", "IGP-DI", "IGP-M",
    "INPC", "IPC", "IPC-FIPE", "IPC-M", "IPC-R", "IPCA", "PÓS", "PRÉ", "SELIC",
    "SEM ÍNDICE", "TBF", "TJLP", "TR", "TR-REAL", "UFIR", "US$ COMERCIAL"),
  start_date = Sys.Date() - 5,
  end_date = Sys.Date()
)
}
\arguments{
\item{indexer}{The indexer for the desired storage data. Defaults to various financial indexers.}

\item{start_date}{The start date for retrieving storage data. Defaults to 5 days ago from the current date.}

\item{end_date}{The end date for retrieving storage data. Defaults to the current date.}
}
\value{
A data frame containing the storage data for the specified indexer and date range.
}
\description{
This function retrieves storage data from the Debentures website for a specified indexer and date range.
}
\examples{
# Retrieve storage data for a specific indexer and date range
storage_data <- get_storage_by_indexer(indexer = "IPCA", start_date = "2022-01-01", end_date = "2022-12-31")

}
\keyword{data}
\keyword{debentures}
\keyword{finance}
\keyword{indexer}
\keyword{storage}
