% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraper.R
\name{get_storage_by_value_adjustment}
\alias{get_storage_by_value_adjustment}
\title{Get Storage Data by Value Adjustment Type and Date Range}
\usage{
get_storage_by_value_adjustment(
  adjustment = "inflation",
  start_date = Sys.Date() - 5,
  end_date = Sys.Date()
)
}
\arguments{
\item{adjustment}{The adjustment type for the desired storage data. Either "inflation" or "interest".}

\item{start_date}{The start date for retrieving storage data. Defaults to 5 days ago from the current date.}

\item{end_date}{The end date for retrieving storage data. Defaults to the current date.}
}
\value{
A data frame containing the storage data for the specified adjustment type and date range.
}
\description{
This function retrieves storage data from the debentures.com.br website for a specified value adjustment type and date range.
}
\examples{
# Retrieve storage data for a specific adjustment type and date range
storage_data <- get_storage_by_value_adjustment(adjustment = "inflation", start_date = "2022-01-01", end_date = "2022-12-31")

}
\keyword{adjustment}
\keyword{data}
\keyword{debentures}
\keyword{finance}
\keyword{storage}
\keyword{value}
