% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraper.R
\name{get_trades_by_cetip_code}
\alias{get_trades_by_cetip_code}
\title{Get Trades Data by CETIP Code and Date Range}
\usage{
get_trades_by_cetip_code(
  cetip_code,
  start_date = Sys.Date() - 1,
  end_date = Sys.Date()
)
}
\arguments{
\item{cetip_code}{The CETIP code for the desired trades data.}

\item{start_date}{The start date for retrieving trades data. Defaults to 1 day ago from the current date.}

\item{end_date}{The end date for retrieving trades data. Defaults to the current date.}
}
\value{
A data frame containing the trades data for the specified CETIP code and date range.
}
\description{
This function retrieves trades data from the Debentures website for a specified CETIP code and date range.
}
\examples{
# Retrieve trades data for a specific CETIP code and date range
trades_data <- get_trades_by_cetip_code(cetip_code = "ABC123", start_date = "2022-01-01", end_date = "2022-12-31")

}
\keyword{CETIP}
\keyword{code}
\keyword{data}
\keyword{debentures}
\keyword{finance}
\keyword{trades}
