% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraper.R
\name{get_trades_by_date}
\alias{get_trades_by_date}
\title{Get Trades Data by Date Range}
\usage{
get_trades_by_date(start_date = Sys.Date() - 5, end_date = Sys.Date())
}
\arguments{
\item{start_date}{The start date for retrieving trades data. Defaults to 5 days ago from the current date.}

\item{end_date}{The end date for retrieving trades data. Defaults to the current date.}
}
\value{
A data frame containing the trades data for the specified date range.
}
\description{
This function retrieves trades data from the Debentures website for a specified date range.
}
\examples{
# Retrieve trades data for the last week
trades_data <- get_trades_by_date()

# Retrieve trades data for a specific date range
trades_data <- get_trades_by_date(start_date = "2022-01-01", end_date = "2022-12-31")

}
\keyword{data}
\keyword{debentures}
\keyword{finance}
\keyword{trades}
