% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{request_data}
\alias{request_data}
\title{Request data from a given URL}
\usage{
request_data(url)
}
\arguments{
\item{url}{The URL to request data from}
}
\value{
The HTTP response object
}
\description{
This function sends an HTTP GET request to the specified URL and returns the response.
}
\examples{
\dontrun{
# Example usage:
url <- "https://example.com/data"
response <- request_data(url)
}
}
